[TMA-06] - Add Filtering, logger.

+++ src/main/java/craftsoft/taskmanagementapi/dao/SubTaskRepository.java
@@ -3,5 +3,5 @@ package craftsoft.taskmanagementapi.dao;
 import craftsoft.taskmanagementapi.domain.SubTask;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SubTaskRepository extends JpaRepository<SubTask, Long> {
+public interface SubTaskRepository extends JpaRepository<SubTask, Integer> {
 }
